Updated MeasureGPMDP.inc to match
MeasureGPMDP now matches the layout of the other Measure#PlayerType#.inc
files. Also fixed where it was monitoring state instead of status.

I also modified mediaplayer.ini to include the option for GPMDP as a
placeholder until it gets rewritten.

I also downscaled both the spotify and gpm logos as they were eating up
huge amounts of CPU and did not need to be that large.