Included installer and helper, fixed control bug
Fixed a bug where Measure state was not included in MeasureGPMDP, I did
not realize that all the controls had to be on that measure name as they
were defined in another file.

I also included the installer and helper for the GPMDP, I dont know if
we will want to change these around/simplify them but if we are
including the .dll file it is not safe to assume that the user has done
the installation steps. Going to see if I can get them to come up if the
user trys to use a media control for GPMDP without having remote
authorization