Performance improvements in settings
Since LeftMouseDownAction fires right away all the settings can just use
an Update of -1. General and Visualization however use the mouse plugin
and can not so their update rates are unchanged for now.

I also changed the encoding on the general settings to ANSI since the °
was encoded incorrectly and was showing up bad in rainmeter.