Color selection performance enhancements
How the color is set has been completely rewritten to remove the need
for DynamicVariables=1 and to get rid of having a bunch of different
styles and groups. Now instead you just set the color directly on the
meter using either the #Color# or #TextColor# variables and then add the
meter to the GroupDynamicColors group. The meter then automatically gets
the now color once it has changed.

This update has two issues where DynamicVariables=1 had to be left and I
am not sure why:
The visualizer bars do not want to change color after the first color
change.
The progress bar does not show up until the first song change.
To fix these bugs I have left dynamic variables on even though it should
not be needed. I will try to remove the in a future commit which should
hopefully drop CPU usage even more.

Typical CPU usage on my i5: 0.2% to 0.4%