Visualizer bars now use new update system
Group was being manually overriden in the rotate script prevent them
from getting updates. Now they use the new color systems push based
variable updates.
Also added the GroupProgressBar group back onto the progress bar so it
is hidden when doing rotations. I have not found a solution for it being
hidden on the first song so it still uses DynamicVariables=1 so it works
correctly.