Even better mouse handling in visualization settings
Now the biggest issue is if you mouse down on one and then mouse over
the next it will start registering you as dragging on it. Hoping I can
avoid a single mouse measure solution for easier coding, going to next
add a check to mouseover similar to the mouse down.