Added SetPosition mouse support
Seems to be working almost flawlessly on my end. Note that spotify does
not yet support seeking but support for that is coming so I still send
the SetProgress command even when using spotify.

The only can only happen if you have settings open at the same time you
are setting position and mouse down on the position and then mouse over
settings while the mouse is still down you can change the settings but
the only person who is gonna do that is me testing edge cases and it can
not be fixed.