Added condition for spotify and bug fixes
Fixed a bug where spotiy's internal double it gives is form 0-1 instead
of the 0-100 everything else is.
Fixed a bug where mouse measure would be disabled if mouse stayed over
the progress bar after click.
Now ignored spotify set positions, this code can easily be changed to
support it when the time comes.